#123 feat: 미디어 메타데이터 API 개선 및 위치 정보 기능 추가
미디어 메타데이터 API 경로 및 DTO 변경
- 기존 경로를 '/v1/trips/{tripId}/media-files'로 수정
- MediaFileRequestDTO를 UpdateMediaFileInfoRequestDTO로 이름 변경
- 새로운 DTO UpdateMediaFileLocationRequestDTO 추가

위치 정보 없는 이미지 처리 기능 추가
- Redis에서 위치 정보 없는 이미지 조회 API 추가
- 위치 정보 업데이트 API 구현

보안 설정 업데이트
- '/v1/**' 경로에 대한 접근 허용 추가
