+++ src/main/java/com/fivefeeling/memory/domain/media/dto/UpdateMediaFileInfoRequestDTO.java
@@ -1,0 +1,14 @@
+package com.fivefeeling.memory.domain.media.dto;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+public record UpdateMediaFileInfoRequestDTO(
+        String mediaLink,
+        Double latitude,
+        Double longitude,
+        String recordDate,
+        String mediaType
+) {
+
+}
