+++ src/main/java/com/fivefeeling/memory/domain/media/dto/UpdateMediaFileLocationRequestDTO.java
@@ -1,0 +1,8 @@
+package com.fivefeeling.memory.domain.media.dto;
+
+public record UpdateMediaFileLocationRequestDTO(
+        Double latitude,
+        Double longitude
+) {
+
+}
